冷却时间文本支持浮点数展示

+++ src/main/java/com/eeeab/eeeabsmobs/sever/util/EMTUtils.java
@@ -76,7 +76,7 @@ public class EMTUtils {
         return component;
     }
 
-    public static Component itemCoolTime(int args) {
+    public static Component itemCoolTime(double args) {
         return simpleText(TOOLTIP_OTHER_PREFIX, args == 0 ? "no_cd" : "cd", STYLE_GREEN, args);
     }
